+++ lib/slack/slack-seen-cache.ts
@@ -40,8 +40,13 @@ export class FlumeSlackSeenCache {
 
     if (this.seen.size <= this.props.maxSize) return
 
-    const entries = [...this.seen.entries()]
-    this.seen = new Map(entries.slice(entries.length - this.props.maxSize))
+    let removeCount = this.seen.size - this.props.maxSize
+
+    for (const id of this.seen.keys()) {
+      if (removeCount <= 0) break
+      this.seen.delete(id)
+      removeCount--
+    }
   }
 
   get size(): number {